backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-pptx==0.6.23
python-multipart==0.0.6
openai==1.12.0
pydantic==2.5.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
slowapi==0.1.9
python-dotenv==1.0.0
xlsxwriter==3.1.9
